refactor(playground-ui): split sidebar refresh boundary
Move MainSidebar context and hooks into a non-component module so the provider remains a valid Fast Refresh boundary without lint suppressions.
